RpiWebServer-master/requirements.txt
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
Flask==1.1.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mysql-connector==2.2.9
paho-mqtt==1.5.0
requests==2.23.0
urllib3==1.25.8
Werkzeug==1.0.0
astroid==2.3.3
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
Flask==1.1.1
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector==2.2.9
paho-mqtt==1.5.0
pylint==2.4.4
requests==2.23.0
six==1.14.0
typed-ast==1.4.1
urllib3==1.25.8
Werkzeug==1.0.0
wrapt==1.11.2
